requirements.txt
streamlit==1.36.0
pandas>=2.1
numpy>=1.26,<2
scikit-learn>=1.3
joblib>=1.3
openpyxl>=3.1
rdkit-pypi>=2022.9.5
